Model: Use size_t rows and append via insert_row in C++
remove_row and insert_row take the row as size_t, like the other
functions of the Model class, and push_row appends with insert_row by
default so that implementing insert_row is enough to support push. The
VectorModel push_row override is no longer needed.

The interpreter vtable uses usize rows like its other row parameters.
The model bridges keep int rows like their sibling functions, and the
array function helpers called by the generated code use int, the type
of a .slint int expression. model_length returns int so the narrowing
from row_count's size_t happens once in the helper.

+++ internal/interpreter/ffi.rs
@@ -610,8 +610,8 @@ pub struct ModelAdaptorVTable {
     pub row_data: unsafe extern "C" fn(VRef<ModelAdaptorVTable>, row: usize) -> *mut Value,
     pub set_row_data: extern "C" fn(VRef<ModelAdaptorVTable>, row: usize, value: Box<Value>),
     pub push_row: extern "C" fn(VRef<ModelAdaptorVTable>, value: Box<Value>),
-    pub remove_row: extern "C" fn(VRef<ModelAdaptorVTable>, row: isize),
-    pub insert_row: extern "C" fn(VRef<ModelAdaptorVTable>, row: isize, value: Box<Value>),
+    pub remove_row: extern "C" fn(VRef<ModelAdaptorVTable>, row: usize),
+    pub insert_row: extern "C" fn(VRef<ModelAdaptorVTable>, row: usize, value: Box<Value>),
     pub get_notify: extern "C" fn(VRef<'_, ModelAdaptorVTable>) -> &ModelNotifyOpaque,
     pub drop: extern "C" fn(VRefMut<ModelAdaptorVTable>),
 }
@@ -644,12 +644,12 @@ impl Model for ModelAdaptorWrapper {
     }
 
     fn remove_row(&self, row: usize) {
-        self.0.remove_row(row as isize);
+        self.0.remove_row(row);
     }
 
     fn insert_row(&self, row: usize, data: Value) {
         let val = Box::new(data);
-        self.0.insert_row(row as isize, val);
+        self.0.insert_row(row, val);
     }
 
     fn as_any(&self) -> &dyn core::any::Any {
